requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.7.2
    # via django
django==4.2.6
    # via
    #   -r requirements.in
    #   django-filter
    #   djangorestframework
django-filter==23.3
    # via -r requirements.in
django-mssql-backend==2.8.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
pyodbc==5.0.1
    # via django-mssql-backend
python-dotenv==1.0.0
    # via -r requirements.in
pytz==2023.3.post1
    # via djangorestframework
sqlparse==0.4.4
    # via django
